+++ cs_interface/cpp/Gyroscope.h
@@ -9,14 +9,14 @@ class Gyroscope
 public:
 	float gain[3];
 	float x, y, z;
-	float raw_x, raw_y, raw_z;
+	int raw_x, raw_y, raw_z;
 	int offset[3];
 	float polarity[3];
 
 	Gyroscope();
 
 	void zeroCalibrate(unsigned int totSamples, unsigned int sampleDelayMS);
-	void updateGyro(float gyroX, float gyroY, float gyroZ);
+	void updateGyro(int gyroX, int gyroY, int gyroZ);
 	void setRevPolarity(bool Xpol, bool Ypol, bool Zpol);
 };
 #endif
